More recognition, and implemented some ROTADA required functions new splits
