+++ configs/screens/sym.saveload.txt
@@ -1,15 +1,21 @@
 // Data
+g_StageStrings                    = 0x801E74A8;
+g_LoadingMemCardTimer             = 0x801E7518;
 g_MemCardState                    = 0x801E751C; // type:s32
 g_MemCardStateTextTimer           = 0x801E7520;
-g_DisplaySaveDataInfo             = 0x801E76D0;
 g_GameState_SaveScreen_Funcs      = 0x801E7524;
 g_IsSaveSelected                  = 0x801E7540;
 g_GameState_DeathLoadScreen_Funcs = 0x801E7544;
-g_StageStrings                    = 0x801E74A8;
 g_OverwriteOptionSelected         = 0x801E755C;
+g_HiddentElementByDisplacement    = 0x801E7570;
+g_DisplaySaveDataInfo             = 0x801E76D0;
+g_IsMemCardNotInserted            = 0x801E76D2;
 g_LastSaveIdx                     = 0x801E76CE;
+g_Gfx_SaveShineTimer              = 0x801E76D4;
+g_IsGameSaving                    = 0x801E76D6;
 
 // Funcs
+Savegame_ScreenSubInit           = 0x801E2D8C; // type:func
 Gfx_SaveBackgroundDraw           = 0x801E2EBC; // type:func
 Gfx_SaveSelectedDisplacement     = 0x801E2F90; // type:func
 Gfx_SaveFileSelectedDraw         = 0x801E2FCC; // type:func
