+++ configs/sym.bodyprog.txt
@@ -18,11 +18,14 @@ g_MapEventIdx = 0x800A9A14; // type:u32
 g_SysState_GameOver_TipIdx = 0x800BCD80; // type:u8
 g_MapEventParam = 0x800BCDD8; // type:u32 - pointer to s_ShEventParam
 
-g_SaveCount = 0x800BCD3C;
-g_SaveSelectedIdx = 0x800A97D4;
+g_SlotElementsCount = 0x800BCD3C;
+g_SlotElementSelectedIdx = 0x800A97D4;
 g_SlotSelectedIdx = 0x800A97D6;
 g_SelectedFileIdx = 0x800BCD3F;
 g_SelectedSaveIdx = 0x800BCD3E;
+g_SlotsElementsInfo = 0x800BCD2C;
+g_MemCardElementCount = 0x800BCD30;
+g_MemCardsTotalElementCount = 0x800BCD3A;
 
 g_IntervalVBlanks = 0x800A8FF0; // type:s32
 g_PrevVBlanks = 0x800A9770; // type:s32
