+++ include/bodyprog/credits_system.h
@@ -1,0 +1,24 @@
+#ifndef _CREDITS_SYSTEM_H
+#define _CREDITS_SYSTEM_H
+
+/** @brief
+* This file is used for define any variable, struct and
+* function that is part of `BODYPROG.BIN` and has been
+* identified to be related to the credits system.
+*/
+
+// ==========
+// FUNCTIONS
+// ==========
+
+void func_80091380();
+
+void func_80091464();
+
+void GameFs_StfRollBinLoad();
+
+void func_800917C0(void*, s32, s_FsImageDesc*);
+
+void func_8009185C(s16 arg0, s16 arg1, s32 arg2, s32 arg3, s32 arg4, s32 arg5, s32 arg6, s32 arg7, s32 arg8, s32 arg9);
+
+#endif
