+++ include/screens/saveload/saveload.h
@@ -11,6 +11,31 @@
 * function that is part of `SAVELOAD.BIN`.
 */
 
+// ============
+// ENUMERATORS
+// ============
+
+
+/** @brief
+* Used by `s_SaveSlotElementInfo` to identify
+* the type of element in the slot save.
+*/
+typedef enum _SlotElementType
+{
+    ElementType_NoMemCard     = 0,
+    ElementType_Unk1          = 1,
+    ElementType_BrokenMemCard = 2,
+    ElementType_LoadMemCard   = 3,
+    ElementType_Unk4          = 4,
+    ElementType_Unk5          = 5,
+    ElementType_Unk6          = 6,
+    ElementType_BrokenSave    = 7,
+    ElementType_Save          = 8,
+    ElementType_NewSave       = 9,
+    ElementType_NewFile       = 10
+} e_SlotElementType;
+
+
 // ================
 // UNKNOWN STRUCTS
 // ================
@@ -19,14 +44,14 @@ that haven't been recognized */
 
 typedef struct
 {
-    s32 unk_0; // It behave the same as field_0 in `s_UnkSaveload0`.
+    s32 unk_0; // Same behaviour as field_0 in `s_SaveSlotElementInfo`.
     u32 gameplayTimer_4;
     u16 savegameCount_8;
     s8  mapEventIdx_A;
     u8  isTitleYellowFlag_B_0 : 1;
     u8  add290Hours_B_1       : 2;
     u8  hyperBlasterFlags_B_3 : 5;
-} s_UnkSaveload1; // Size assumed.
+} s_SaveBasicInfo;
 /** Possible size is 0xC/12 bytes. */
 
 typedef struct
@@ -48,15 +73,15 @@ typedef struct
 	* card and then saves in slot 1 the value won't
 	* be 3, but instead it will be 2.
 	*/
-    s16             field_0;
-    s16             savegameCount_2;
-    s8              field_4; // Normally the value it set as 8 unless it is the `New save` option where it is set as 9.
-    s8              field_5; // The value changes between 0 when the first save slot is selected and 4 when the second is selected.
-    s8              file_Idx_6; // Also counts the `New save` space.
-    s8              save_Idx_7;
-    s32             field_8; // Maybe bitfield.
-    s_UnkSaveload1* field_C;
-} s_UnkSaveload0; // Size: >=12
+    s16              field_0;
+    s16              savegameCount_2;
+    s8               elementType_4; /** `e_SlotElementType` */
+    s8               field_5; // The value changes between 0 when the first save slot is selected and 4 when the second is selected.
+    s8               file_Idx_6;
+    s8               element_Idx_7;
+    s32              mapEventIdx_8;
+    s_SaveBasicInfo* field_C;
+} s_SaveSlotElementInfo;
 /**
 * Possible size is 0x10/16 bytes.
 * This struct is written in memory each time a
@@ -142,7 +167,9 @@ extern char* g_StageStrings[];
 * he confirms the overwrite then `D_801E750C`
 * turns 0 and `D_801E7510` 1, but if the
 * player cancels the overwrite then they keep
-* the values (`D_801E750C` 1 and `D_801E7510` 0)
+* the values (`D_801E750C` 1 and `D_801E7510` 0).
+*
+* `D_801E7510` it is also related to loading.
 */
 extern s32 D_801E750C;
 
@@ -150,7 +177,7 @@ extern s32 D_801E7510;
 
 extern s16 D_801E7514[MEMORY_CARD_SLOT_COUNT];
 
-extern s16 D_801E7518[MEMORY_CARD_SLOT_COUNT]; // Timer that triggers when loading a detected memory card
+extern s16 g_LoadingMemCardTimer[MEMORY_CARD_SLOT_COUNT]; // Timer that triggers when loading a detected memory card.
 
 extern s32 g_MemCardState; // Format: 1, Save: 2, Load: 3
 
@@ -185,7 +212,11 @@ extern s32 D_801E7564[MEMORY_CARD_SLOT_COUNT];
 */
 extern s16 D_801E756C[MEMORY_CARD_SLOT_COUNT];
 
-extern s16 D_801E7570[MEMORY_CARD_SLOT_COUNT];
+/** @brief
+* Count elements that get hidden as the player move
+* down the slot.
+*/
+extern s16 g_HiddentElementByDisplacement[MEMORY_CARD_SLOT_COUNT];
 
 extern s16 D_801E7574[MEMORY_CARD_SLOT_COUNT];
 
@@ -193,31 +224,25 @@ extern s16 D_801E7574[MEMORY_CARD_SLOT_COUNT];
 * Stores the selected save in each slot.
 * Only the first two values are used.
 * Possibly the next values are an unused
-* which Splat detect incorrectly and merge
-* it.
+* variable which Splat detect incorrectly
+* and merge it.
 */
-extern s16 D_801E7578[6];
+extern s16 D_801E7578[MEMORY_CARD_SLOT_COUNT];
 
-extern s8 D_801E7584[330];
+extern s8 D_801E7584[SAVE_COUNT_MAX * 2];
 
 extern s8 g_LastSaveIdx[MEMORY_CARD_SLOT_COUNT];
 
 extern s8 g_DisplaySaveDataInfo;
 
-/** Seems to check if the memory cards are
-* loaded, if they aren't then the value
-* of each is 1, otherwise it is 0.
-*/
-extern s8 D_801E76D2[MEMORY_CARD_SLOT_COUNT];
+extern s8 g_IsMemCardNotInserted[MEMORY_CARD_SLOT_COUNT]; // 0 - is inserted, 1 - is not inserted.
 
-// Timer used for the shine animation after saving a game.
-extern s8 D_801E76D4;
+extern s8 g_Gfx_SaveShineTimer;
 
 // Use when overwriting a save with yellow text.
 extern u8 D_801E76D5;
 
-// Is game saving flag?
-extern u8 D_801E76D6;
+extern u8 g_IsGameSaving;
 
 
 // ==========
@@ -226,7 +251,7 @@ extern u8 D_801E76D6;
 
 
 /** Clears all saves? */
-void func_801E2D8C();
+void Savegame_ScreenSubInit();
 
 void Gfx_SaveSelectedDisplacement(s32 saveSlotIdx, s32 arg1);
 
@@ -234,12 +259,12 @@ void Gfx_SaveSelectedDisplacement(s32 saveSlotIdx, s32 arg1);
 void Gfx_SaveFileSelectedDraw(s32 arg0, s32 saveSlotIdx, s32 fileId, s32 arg3);
 
 /** Draws the string of the location where the save was done. */
-void Gfx_SavesLocationDraw(s_UnkSaveload0* ptr, s32 arg1, s32 idx);
+void Gfx_SavesLocationDraw(s_SaveSlotElementInfo* ptr, s32 arg1, s32 idx);
 
 /** Sets the color of the string to be drawn based on some flag. */
-s32 func_801E3078(s_UnkSaveload1* arg0);
+s32 func_801E3078(s_SaveBasicInfo* arg0);
 
-void func_801E326C(s_UnkSaveload0* arg0, s_UnkSaveload0* arg1, s32 arg2, s32 arg3);
+void func_801E326C(s_SaveSlotElementInfo* arg0, s_SaveSlotElementInfo* arg1, s32 arg2, s32 arg3);
 
 void Gfx_MemCardStateDraw(s32 memCardState, s32 arg1);
